Changed deploy.sh

+++ AutoFindBot.Web/deploy.sh
@@ -2,8 +2,30 @@
 
 set -e # включение проверки ошибок
 
+# Обработка аргументов командной строки
+OPTIONS=$(getopt -o t: --long token: -n 'parse-options' -- "$@")
+eval set -- "$OPTIONS"
+
+TOKEN=""
+
+while true; do
+  case "$1" in
+    -t|--token) TOKEN="$2"; shift 2;;
+    --) shift; break;;
+    *) echo "Ошибка: некорректный аргумент $1" >&2; exit 1;;
+  esac
+done
+
+if [ -z "$TOKEN" ]; then
+  echo "Ошибка: токен не указан" >&2
+  exit 1
+fi
+
 echo "Удаление существующего образа..."
 docker rmi -f autofindbot >/dev/null 2>&1 || true # игнорирование ошибки, если образ уже удален
+
+echo "Клонирование репозитория из GitHub..."
+git clone https://${TOKEN}@github.com/max-tkv/AutoFindBot.git /home/dev/AutoFindBot
 
 echo "Создание нового образа..."
 docker build -f "/home/dev/AutoFindBot/AutoFindBot.Web/Dockerfile" --force-rm -t autofindbot "/home/dev/AutoFindBot" || { echo "Ошибка: не удалось создать образ" >&2; exit 1; }
